% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensembles}
\alias{ensembles}
\title{Ensemble models based on a random forest architecture that pools the predictions of six labeling systems generated using \code{detect_sdg_systems()} and also considers text length.}
\format{
An object of class \code{list} of length 4.
}
\source{
Wulff, Meier, & Mata (2023)
}
\usage{
ensembles
}
\description{
A list of trained \code{ranger::ranger()} random forest models that are used by the text2sdg \code{detect_sdg()} function.
}
\keyword{datasets}
