% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensembles}
\alias{ensembles}
\title{A list of trained \code{ranger::ranger()} random forest models that are used by the text2sdg \code{detect_sdg()} function.}
\format{
An object of class \code{list} of length 4.
}
\source{
Wulff, D. U., Meier, D., & Mata, R. (2023). Using novel data and ensemble models to improve automated SDG-labeling. arXiv
}
\usage{
ensembles
}
\description{
Ensemble models based on a random forest architecture that pools the predictions of six labeling systems generated using the \code{detect_sdg_systems()} function from the text2sdg package and also considers text length.
}
\keyword{datasets}
